requirements.txt
Django==2.1.3
django-redis==4.10.0
djangorestframework==3.9.0
ipython==7.2.0
mysqlclient==1.3.13
qiniu==7.2.2
celery==4.2.1
django-celery-results==1.0.4
uWSGI==2.0.17.1
gunicorn==19.9.0
gevent==1.3.7
django-filter==2.1.0
django-crispy-forms==1.7.2


